AWS/Lesson-8/pre-built/old-setup/requirements.txt
Flask==2.1.1
mysql-connector-python==8.0.28
Werkzeug==2.0.3
